Recipes/requirements.txt
django
django-ninja
psycopg2-binary
Pillow